+++ pkg/worker/worker.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/iver-wharf/wharf-cmd/pkg/wharfyml"
+	"github.com/iver-wharf/wharf-cmd/pkg/worker/workermodel"
 	"github.com/iver-wharf/wharf-core/pkg/logger"
 )
 
@@ -34,7 +35,7 @@ type StageRunner interface {
 
 // StageRunnerFactory creates a new StageRunner for a given stage.
 type StageRunnerFactory interface {
-	NewStageRunner(ctx context.Context, stage wharfyml.Stage) (StageRunner, error)
+	NewStageRunner(ctx context.Context, stage wharfyml.Stage, stepIDOffset uint64) (StageRunner, error)
 }
 
 // StepRunner is the interface for running a Wharf build step. Steps are the
@@ -47,35 +48,35 @@ type StepRunner interface {
 
 // StepRunnerFactory creates a new StepRunner for a given step.
 type StepRunnerFactory interface {
-	NewStepRunner(ctx context.Context, step wharfyml.Step) (StepRunner, error)
+	NewStepRunner(ctx context.Context, step wharfyml.Step, stepID uint64) (StepRunner, error)
 }
 
 // Result is a Wharf build result with the overall status of all stages were
 // executed, the induvidual stage results, as well as the duration of the entire
 // Wharf build.
 type Result struct {
-	Status   Status        // execution status of the entire build
-	Options  BuildOptions  // options used when running the build
-	Stages   []StageResult // execution results for each stage
-	Duration time.Duration // execution duration of the entire build
+	Status   workermodel.Status // execution status of the entire build
+	Options  BuildOptions       // options used when running the build
+	Stages   []StageResult      // execution results for each stage
+	Duration time.Duration      // execution duration of the entire build
 }
 
 // StageResult is a Wharf build stage result with the overall status of the
 // steps that was executed for the stage, as well as the duration of the
 // Wharf build stage.
 type StageResult struct {
-	Name     string        // name of the stage
-	Status   Status        // execution status of the stage
-	Steps    []StepResult  // execution results for each step
-	Duration time.Duration // execution duration of the stage
+	Name     string             // name of the stage
+	Status   workermodel.Status // execution status of the stage
+	Steps    []StepResult       // execution results for each step
+	Duration time.Duration      // execution duration of the stage
 }
 
 // StepResult is a Wharf build step result with the status of the step execution
 // as well as the duration of the Wharf build step.
 type StepResult struct {
-	Name     string        // name of the step
-	Status   Status        // execution status of the step
-	Type     string        // type of Wharf build step, eg. "container" or "docker"
-	Error    error         // error message from the execution, if any
-	Duration time.Duration // execution duration of the step
+	Name     string             // name of the step
+	Status   workermodel.Status // execution status of the step
+	Type     string             // type of Wharf build step, eg. "container" or "docker"
+	Error    error              // error message from the execution, if any
+	Duration time.Duration      // execution duration of the step
 }